Update install_latest_blue_onyx.ps1 (#36)
* Update install_latest_blue_onyx.ps1

Updated install options to include a popup folder selection to install the app.

Added 4th bat file for Large models.

Updated final color script to break up the Green Noise. 

Added user inputbox for GPU selection, the value of that selection then gets put into the BAT files. 

Get a list of available GPUs 

We also want to exclude "RDP Displays" and Sort the list like how it is in Task Manager so the Index is correct. 

Added Logic to compare the installed version to the JSON on github to determine if installing a newer version is needed or not. 

Cleaned up step formatting.

Added Popup box for the end user to select where to install the PATH variable. If Self Only then the Variable gets added to User if System Wide then the Variable gets added to the System. 

If Multiple GPU's detected then present popup selection box, else default value to 0

added GPUS to an Array logic